Extract hook to measure header height
